delete_image authorization

+++ server.py
@@ -180,17 +180,31 @@ def add_image_a(question_id, answer_id):
 
 @app.route('/question/<int:question_id>/delete-image')
 def delete_question_image(question_id):
-    filename = 'static/image_for_question' + str(question_id) + '.png'
-    util.delete_file(filename)
-    data_manager.update_image_data_by_id(data_manager.question_db, question_id, 'no image')
+    try:
+        user_id = data_manager.get_userid_by_username(session['username'])
+    except KeyError:
+        return redirect('/')
+    if user_id == data_manager.get_foreign_key_by_id(data_manager.answer_db, 'users_id', answer_id)[0]['users_id']:
+        filename = 'static/image_for_question' + str(question_id) + '.png'
+        util.delete_file(filename)
+        data_manager.update_image_data_by_id(data_manager.question_db, question_id, 'no image')
+    else:
+        flash('Invalid user')
     return redirect('/question/{}'.format(question_id))
 
 
 @app.route('/question/<int:question_id>/<int:answer_id>/delete-image')
 def delete_answer_image(question_id, answer_id):
-    filename = 'static/image_for_answer' + str(answer_id) + '.png'
-    util.delete_file(filename)
-    data_manager.update_image_data_by_id(data_manager.answer_db, answer_id, 'no image')
+    try:
+        user_id = data_manager.get_userid_by_username(session['username'])
+    except KeyError:
+        return redirect('/')
+    if user_id == data_manager.get_foreign_key_by_id(data_manager.answer_db, 'users_id', answer_id)[0]['users_id']:
+        filename = 'static/image_for_answer' + str(answer_id) + '.png'
+        util.delete_file(filename)
+        data_manager.update_image_data_by_id(data_manager.answer_db, answer_id, 'no image')
+    else:
+        flash('Invalid user')
     return redirect('/question/{}'.format(question_id))
